API helper proposal (#111)
* API helper proposal

* login uses new helpers

* remove unnecessary type definitions

* Nullable jwt
