+++ src/tapis-api/src/utils/index.ts
@@ -1,1 +1,3 @@
-export { default as queryHelper } from './queryHelper';
+
+export { default as apiGenerator } from './apiGenerator';
+export { default as errorDecoder } from './errorDecoder';
